requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
certifi==2020.12.5
flake8==3.8.4
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
